Completed test() and added eval during training
The test function can be used to evaluate a given model. It has the
option to be used to render the environment as well. Added eval
during training using the eval callback from stable_baselines3. A
benchmark_random_agent() function has also been added to benchmark
the performance of a random agent.